+++ Lab-kotlin/src/main/kotlin/com/petclinic/core/OwnerRepository.kt
@@ -2,6 +2,6 @@ package com.petclinic.core
 
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface OwnerRepository: JpaRepository<Owner, Int> {
+interface OwnerRepository: JpaRepository<Owner, Long> {
     fun findByFirstName(firstName: String): Owner?
 }
