+++ Lab-kotlin/src/main/kotlin/com/petclinic/core/VisitRepository.kt
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Component
 
 @Component
-interface VisitRepository: JpaRepository<Visit, Int> {
+interface VisitRepository: JpaRepository<Visit, Long> {
 }
